created a new valueTransformer for TransformValue()
